Add files via upload
Develop a C# app that will determine the gross pay for each of three employees. The company pays
straight time for the first 20 hours worked by each employee and time-and-a-half for all hours worked
in excess of 20 hours. You’re given a list of the three employees of the company, the number of hours
each employee worked last week and the hourly rate of each employee. Your app should input this
information for each employee, then should determine and display the employee’s gross pay. Use the
Console class’s ReadLine method to input the data.